requirements.txt
pillow==8.2.0
numpy==1.19.2
pandas==1.1.3
scipy==1.5.2
matplotlib==3.3.2
